Move ClearCommand from exec to store
The core interpreter should have zero command implementations in it, and so
ClearCommand does not belong in there.  Move it from exec to store.

The new home within store.rs isn't the greatest fit given that CLEAR
does not touch the stored program, but is the best fit at the moment
given that adding a new module just for this doesn't seem worthwhile.
